Embedding Bag Primitive Addition - Added an embedding bag primitive. This primitive mirrors the   Pytorch nn.EmbeddingBag operator. - Implemented reference kernel that works with f32 embedding   tables, and s32 indices and offsets. - Implemented gtests for the new primitive. - Implemented benchdnn driver for the new primitive

+++ examples/primitives/embedding_bag.cpp
@@ -135,7 +135,7 @@ void embedding_bag_example(dnnl::engine::kind engine_kind) {
     read_from_dnnl_memory(dst_data.data(), dst_mem);
 
     // Check correctness
-    for (int i = 0; i < dst_data.size(); ++i)
+    for (uint64_t i = 0; i < dst_data.size(); ++i)
       std::cout << "i:" << i << ":" << dst_data[i] << std::endl;
 }
